pyproject.toml
[tool.coverage.run]
branch = true
omit = ["config.py", "config-3.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if typing.TYPE_CHECKING:",
]

[tool.mypy]
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
strict_equality = true

[tool.poe.tasks]
deploy = "robotpy deploy"
download = "robotpy sync --no-install"
sim = "robotpy sim"
test = "robotpy test"
type-check = "mypy ."

[tool.pyright]
exclude = [
    "ctre_sim",
    "logs",
    # Default excludes
    "**/__pycache__",
    "**/.*",
]

[tool.pytest.ini_options]
addopts = "--strict-markers -v --maxfail=2"
faulthandler_timeout = 60
pythonpath = "."
testpaths = ["tests"]
xfail_strict = true

[tool.ruff]
target-version = "py311"

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # pyflakes
    "F",
    # flake8-bugbear
    "B",
    # isort
    "I",
    # pyupgrade
    "UP",
    # flake8-comprehensions
    "C4",
    # flake8-logging-format
    "G",
    # flake8-simplify
    "SIM",
    # flake8-print
    "T20",
    # perflint
    "PERF",
    # ruff specific
    "RUF021", # parenthesise chained and/or
]
ignore = [
    "E501", # long lines
    "B905", # zip() without strict=
]

[dependency-groups]
dev = [
    "coverage>=7.10.2",
    "hypothesis>=6.112.1",
    "mypy>=1.19.1",
    "pytest>=7.2.0",
    "pytest-integration>=0.2.3",
]

[tool.uv]
prerelease = "allow"

[project]
name = "pyrebuilt"
version = "0.0.0"
description = "The Warriors' FRC 2026 robot code"
authors = [{ name = "The Warriors of East Harlem" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.14,<3.15"

dependencies = [
    "numpy~=2.3",
    "phoenix6~=26.1.0",
    "robotpy-ctre~=2026.1.0",
    "robotpy[apriltag]==2026.1.1",
    "robotpy-rev==2026.0.0",
    "robotpy-wpilib-utilities~=2026.0.1",
    "photonlibpy==2026.1.1rc3",
    "sleipnirgroup-choreolib~=2026.0.1",
]

[tool.robotpy]
requires = [
    "phoenix6~=26.1.0",
    "robotpy-ctre~=2026.1.0",
    "robotpy-rev==2026.0.0",
    "robotpy-wpilib-utilities==2026.0.1",
    "photonlibpy==2026.1.1rc3",
    "sleipnirgroup-choreolib~=2026.0.1",
]
robotpy_version = "2026.1.1"
components = ["apriltag"]
